Move the bundled robot models out of `roboarm/__init__.py` into `roboarm.models`
`__init__.py` keeps only the package contract (`State`, `RobotStatus`, `command`); the URDF + mesh
model builders and the 2F-85 graft move to `roboarm.models`, and consumers import from there.
