+++ positronic/tests/test_inference_integration.py
@@ -9,7 +9,7 @@ import tqdm
 import positronic.cfg.simulator
 from positronic.cfg.eval.sim.positronic import stack_cubes
 from positronic.dataset.local_dataset import LocalDataset
-from positronic.drivers.roboarm import bundled_panda_model
+from positronic.drivers.roboarm.models import bundled_panda_model
 from positronic.inference import main
 from positronic.policy.tests.test_harness import StubPolicy
 from positronic.simulator.mujoco.sim import MujocoSim
